+++ CMakeLists.txt
@@ -417,17 +417,33 @@ if (BUILD_TESTING)
         file(GLOB BENCHMARK_SRC "tests/benchmark/*.cc")
         enable_language(CXX)
         foreach(benchmark ${BENCHMARK_SRC})
-            string(REGEX REPLACE ".+\\/(.+)\\.cc" "\\1" benchmark_name ${benchmark})
-            add_executable(${benchmark_name} ${benchmark} "bin/echo.c" "bin/common.c" "tests/benchmark/s2n_neg_client_benchmark.cxx" "tests/benchmark/s2n_neg_server_benchmark.cxx")
-            target_include_directories(${benchmark_name} PRIVATE api)
-            target_include_directories(${benchmark_name} PRIVATE tests)
-            target_link_libraries(${benchmark_name} PUBLIC ${PROJECT_NAME} testss2n benchmark::benchmark)
+            if(NOT((benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_negotiate_client_benchmark.cc") OR (benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_negotiate_server_benchmark.cc") OR (benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_negotiate_benchmark.cc") OR (benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_server_benchmark.cc") OR (benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_client_benchmark.cc") OR (benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_benchmark.cc") OR (benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/shared_info.cc")))
+                message(${benchmark})
+                string(REGEX REPLACE ".+\\/(.+)\\.cc" "\\1" benchmark_name ${benchmark})
+                add_executable(${benchmark_name} ${benchmark} "bin/echo.c" "bin/common.c")
+                target_include_directories(${benchmark_name} PRIVATE api)
+                target_include_directories(${benchmark_name} PRIVATE tests)
+                target_link_libraries(${benchmark_name} PUBLIC ${PROJECT_NAME} testss2n benchmark::benchmark)
 
-            # Based off the flags in tests/benchmark/Makefile
-            target_compile_options(${benchmark_name} PRIVATE -pedantic -Wall -Werror -Wunused -Wcomment -Wchar-subscripts
-                    -Wuninitialized -Wshadow -Wcast-qual -Wcast-align -Wwrite-strings -Wno-deprecated-declarations
-                    -Wno-unknown-pragmas -Wformat-security -Wno-missing-braces -fvisibility=hidden -Wno-unreachable-code
-                    -Wno-unused-but-set-variable)
+                # Based off the flags in tests/benchmark/Makefile
+                target_compile_options(${benchmark_name} PRIVATE -pedantic -Wall -Werror -Wunused -Wcomment -Wchar-subscripts
+                        -Wuninitialized -Wshadow -Wcast-qual -Wcast-align -Wwrite-strings -Wno-deprecated-declarations
+                        -Wno-unknown-pragmas -Wformat-security -Wno-missing-braces -fvisibility=hidden -Wno-unreachable-code
+                        -Wno-unused-but-set-variable)
+            elseif(benchmark STREQUAL "/home/ubuntu/s2n-tls/tests/benchmark/s2n_negotiate_benchmark.cc")
+                message(${benchmark})
+                string(REGEX REPLACE ".+\\/(.+)\\.cc" "\\1" benchmark_name ${benchmark})
+                add_executable(${benchmark_name} ${benchmark} "bin/echo.c" "bin/common.c" "tests/benchmark/shared_info.cc" "tests/benchmark/s2n_negotiate_client_benchmark.cc" "tests/benchmark/s2n_negotiate_server_benchmark.cc")
+                target_include_directories(${benchmark_name} PRIVATE api)
+                target_include_directories(${benchmark_name} PRIVATE tests)
+                target_link_libraries(${benchmark_name} PUBLIC ${PROJECT_NAME} testss2n benchmark::benchmark)
+
+                # Based off the flags in tests/benchmark/Makefile
+                target_compile_options(${benchmark_name} PRIVATE -pedantic -Wall -Werror -Wunused -Wcomment -Wchar-subscripts
+                        -Wuninitialized -Wshadow -Wcast-qual -Wcast-align -Wwrite-strings -Wno-deprecated-declarations
+                        -Wno-unknown-pragmas -Wformat-security -Wno-missing-braces -fvisibility=hidden -Wno-unreachable-code
+                        -Wno-unused-but-set-variable)
+            endif()
         endforeach(benchmark)
 
     endif()
