+++ tests/benchmark/server_info.h
@@ -34,3 +34,5 @@ struct conn_settings {
     char *psk_optarg_list[S2N_MAX_PSK_LIST_LENGTH];
     size_t psk_list_len;
 };
+
+
