+++ tests/benchmark/shared_info.h
@@ -14,3 +14,50 @@
  */
 
 
+
+#include <s2n.h>
+#include "tls/s2n_cipher_suites.h"
+#include <benchmark/benchmark.h>
+#include <stdlib.h>
+#include <string.h>
+#include <cstring>
+#include <string>
+
+#include <vector>
+
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <sys/ioctl.h>
+#include <sys/param.h>
+#include <poll.h>
+#include <netdb.h>
+#include <stdlib.h>
+#include <signal.h>
+#include <unistd.h>
+#include <string.h>
+#include <stdio.h>
+#include <getopt.h>
+#include <strings.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <netinet/tcp.h>
+#include <iostream>
+#include <sys/stat.h>
+#include <sys/mman.h>
+#include "utils/s2n_safety.h"
+#include "utils/s2n_random.h"
+
+
+
+
+extern struct s2n_cipher_suite *all_suites[36];
+
+extern uint8_t ticket_key_name[16];
+
+extern uint8_t default_ticket_key[32];
+
+extern int DEBUG_PRINT;
+extern int WARMUP_ITERS;
+extern unsigned int ITERATIONS;
+
+int benchmark_negotiate(struct s2n_connection *conn, int fd, benchmark::State& state, bool warmup);
